removed null files and moved Repository interfaces to the impl package
