+++ src/main/java/za/ac/cput/LibrarySystem/repository/impl/IAuthorRepository.java
@@ -1,4 +1,4 @@
-package za.ac.cput.LibrarySystem.repository;
+package za.ac.cput.LibrarySystem.repository.impl;
 /*
     IAuthorRepository.java
     Interface for Author Repository
@@ -6,9 +6,11 @@ package za.ac.cput.LibrarySystem.repository;
     Date: 9 April 2022
  */
 import za.ac.cput.LibrarySystem.domain.entity.Author;
+import za.ac.cput.LibrarySystem.repository.IRepository;
+
 import java.util.Set;
 
-public interface IAuthorRepository extends IRepository<Author, String>{
+public interface IAuthorRepository extends IRepository<Author, String> {
     Set<Author> getAll();
 
 }
