+++ src/main/java/za/ac/cput/LibrarySystem/repository/impl/IBookRepository.java
@@ -1,4 +1,4 @@
-package za.ac.cput.LibrarySystem.repository;
+package za.ac.cput.LibrarySystem.repository.impl;
 /*
     IBookRepository.java
     Interface for Book Repository
@@ -7,8 +7,10 @@ package za.ac.cput.LibrarySystem.repository;
  */
 
 import za.ac.cput.LibrarySystem.domain.entity.Book;
+import za.ac.cput.LibrarySystem.repository.IRepository;
+
 import java.util.Set;
 
-public interface IBookRepository extends IRepository<Book, String>{
+public interface IBookRepository extends IRepository<Book, String> {
     Set<Book> getAll();
 }
